pyproject.toml
[project]
name = "employee-task-tracker"
version = "0.1.0"
description = ""
authors = [
    {name = "Valdemar",email = "vovateslenko.1996.28@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "django-filter (>=25.1,<26.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "coverage (>=7.8.0,<8.0.0)",
    "django-extensions (>=4.1,<5.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
